+++ plugins/statuspage/dev/__fixtures__/mockGetComponentGroups.json
@@ -1,0 +1,22 @@
+[
+  {
+    "id": "g1",
+    "page_id": "p1",
+    "name": "Backend infrastructure",
+    "description": "Backend services handling events, logging, metrics...",
+    "components": ["c1", "c2", "c3"],
+    "position": 1,
+    "created_at": "2023-03-06T22:00:35Z",
+    "updated_at": "2023-09-11T10:58:01Z"
+  },
+  {
+    "id": "g2",
+    "page_id": "p1",
+    "name": "Backstage",
+    "description": null,
+    "components": ["c4", "c5"],
+    "position": 2,
+    "created_at": "2023-03-27T20:51:07Z",
+    "updated_at": "2023-09-11T10:58:01Z"
+  }
+]
