+++ plugins/statuspage/dev/__fixtures__/mockGetComponents.json
@@ -1,0 +1,114 @@
+[
+  {
+    "id": "g1",
+    "page_id": "p1",
+    "group_id": null,
+    "created_at": "2023-03-06T22:00:35Z",
+    "updated_at": "2023-09-11T10:58:01Z",
+    "group": true,
+    "name": "Backend infrastructure",
+    "description": "Backend services handling events, logging, metrics...",
+    "position": 1,
+    "status": "operational",
+    "showcase": false,
+    "only_show_if_degraded": false,
+    "automation_email": "",
+    "start_date": null
+  },
+  {
+    "id": "c1",
+    "page_id": "p1",
+    "group_id": "g1",
+    "created_at": "2023-03-06T22:00:35Z",
+    "updated_at": "2023-03-29T07:00:31Z",
+    "group": false,
+    "name": "RabbitMQ",
+    "description": "RabbitMQ cluster",
+    "position": 1,
+    "status": "operational",
+    "showcase": true,
+    "only_show_if_degraded": false,
+    "automation_email": "",
+    "start_date": "2023-03-06"
+  },
+  {
+    "id": "c4",
+    "page_id": "p1",
+    "group_id": "g2",
+    "created_at": "2023-03-27T20:51:07Z",
+    "updated_at": "2023-04-12T09:10:43Z",
+    "group": false,
+    "name": "Backstage Pages",
+    "description": "Service for hosting non-techdoc backstage documentation",
+    "position": 1,
+    "status": "operational",
+    "showcase": true,
+    "only_show_if_degraded": false,
+    "automation_email": "",
+    "start_date": "2023-03-27"
+  },
+  {
+    "id": "c5",
+    "page_id": "p1",
+    "group_id": "g2",
+    "created_at": "2023-03-06T09:50:24Z",
+    "updated_at": "2023-04-12T09:10:43Z",
+    "group": false,
+    "name": "Backstage",
+    "description": "Developer Portal. Site served at https://backstage.domain.com/",
+    "position": 2,
+    "status": "operational",
+    "showcase": true,
+    "only_show_if_degraded": false,
+    "automation_email": "",
+    "start_date": "2023-03-06"
+  },
+  {
+    "id": "c2",
+    "page_id": "p1",
+    "group_id": "g1",
+    "created_at": "2023-03-06T22:01:37Z",
+    "updated_at": "2023-03-06T22:03:38Z",
+    "group": false,
+    "name": "GraphQL API to Eiffel event repository",
+    "description": "The GraphQL API for accessing the production Eiffel event repository.",
+    "position": 2,
+    "status": "operational",
+    "showcase": true,
+    "only_show_if_degraded": false,
+    "automation_email": "",
+    "start_date": "2023-03-06"
+  },
+  {
+    "id": "g2",
+    "page_id": "p1",
+    "group_id": null,
+    "created_at": "2023-03-27T20:51:07Z",
+    "updated_at": "2023-09-11T10:58:01Z",
+    "group": true,
+    "name": "Backstage",
+    "description": null,
+    "position": 2,
+    "status": "operational",
+    "showcase": false,
+    "only_show_if_degraded": false,
+    "automation_email": "",
+    "start_date": null
+  },
+  {
+    "id": "c3",
+    "page_id": "p1",
+    "group_id": "g1",
+    "created_at": "2023-03-06T22:04:47Z",
+    "updated_at": "2023-03-06T22:04:47Z",
+    "group": false,
+    "name": "REST API to Eiffel event repository",
+    "description": "The REST API for accessing the production Eiffel event repository.",
+    "position": 3,
+    "status": "operational",
+    "showcase": true,
+    "only_show_if_degraded": false,
+    "automation_email": "",
+    "start_date": "2023-03-06"
+  }
+]
